+++ out/providers/codeLensProvider.js
@@ -25,6 +25,7 @@ var __importStar = (this && this.__importStar) || function (mod) {
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.CodeLensProvider = void 0;
 const vscode = __importStar(require("vscode"));
+const navigateToSectionCSS_1 = require("../commands/navigateToSectionCSS");
 class CodeLensProvider {
     onDidChangeCodeLenses;
     provideCodeLenses(document, token) {
@@ -81,6 +82,19 @@ class CodeLensProvider {
             codeLenses.push(new vscode.CodeLens(range, {
                 title: "Optimize sharp images",
                 command: "codestitchHelper.optimizeSharpImages",
+                arguments: [document, range],
+            }));
+        }
+        // Regex to find <section> tags with an id attribute
+        const sectionPattern = /<section[^>]*id="([^"]+)"[^>]*>/g;
+        while ((match = sectionPattern.exec(text)) !== null) {
+            const startPosition = document.positionAt(match.index);
+            const endPosition = document.positionAt(match.index + match[0].length);
+            const range = new vscode.Range(startPosition, endPosition);
+            // Create a new CodeLens above the section tag
+            codeLenses.push(new vscode.CodeLens(range, {
+                title: "Go to Styling",
+                command: navigateToSectionCSS_1.navigateToSectionCSSCommandId,
                 arguments: [document, range],
             }));
         }