+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0
 - Added optimize images for sharp integration
 - Added button to quickly add code needed for sharp integration
+- Added button to navigate to the styling of a component
 
 ## 0.0.10